backend/requirements.txt
drf-extra-fields==3.7.0
djoser==2.1.0
djangorestframework==3.15.1
Django==5.0.4
django-import-export==3.3.8
pillow==10.3.0
python-dotenv==1.0.1
